Add GeoSeries, GeoDataFrame, DaskGeoSeries, and DaskGeoDataFrame classes
GeoSeries and GeoDataFrame mimic a small portion of the geopandas API,
in particular the `cx` spatial indexer.

The DaskGeoSeries and DashGeoDataFrame support a distributed `cx` that
uses two layers of R-trees. A top-level R-tree is used to identify which
partitions to include and then an R-tree is used in each of the
partitions to complete the `cx` operation.

The GeoDataFrame and DaskGeoDataFrame classes use the geopandas concept
of an active geometry column to control which column `cx` applies to.